refactored code to fix indentation issues and remove code duplication
