+++ app/Http/Controllers/AdvancePaymentController.php
@@ -6,33 +6,14 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Carbon\Carbon;
 use App\Models\Payment; 
-use App\Models\WaterConnection;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
-
 
 class AdvancePaymentController extends Controller
 {
     public function index()
     {
-        $authUser = auth()->user();
-        $now = Carbon::now()->startOfMonth();
-
-        $customers = DB::table('customers')
-            ->join('payments', 'customers.id', '=', 'payments.customer_id')
-            ->join('debts', 'payments.debt_id', '=', 'debts.id')
-            ->where('debts.status', 'paid')
-            ->where('debts.end_date', '>', $now)
-            ->select('customers.id', 'customers.name', 'customers.last_name')
-            ->distinct()
-            ->get();
- 
-        $payments = Payment::with('debt.customer')
-            ->orderBy('id', 'desc')
-            ->paginate(10);
-
-        return view('advancePayments.index', compact('payments', 'customers'));
-       
+        return view('advancePayments.index');
     }
 
     public function create()
@@ -64,37 +45,41 @@ class AdvancePaymentController extends Controller
     {
     
     }
-
-    public function getAdvanceWaterConnections(Request $request)
+    
+    public function getAdvanceCustomersAndConnections(Request $request)
     {
-    $authUser = auth()->user();
-    $customerId = $request->input('customer_id');
+        $currentMonthStart = Carbon::now()->startOfMonth();
+    
+        $paymentsQuery = Payment::with(['customer', 'debt.waterConnection'])
+            ->whereHas('debt', function($debtQuery) use ($currentMonthStart) {
+                $debtQuery->where('status', 'paid')
+                    ->where('end_date', '>', $currentMonthStart);
+            });
 
-    $waterConnections = WaterConnection::where('customer_id', $customerId)
-        ->where('locality_id', $authUser->locality_id)
-        ->whereHas('debts', function ($query) {
-            $query->where('status', 'paid')
-                  ->whereDate('end_date', '>', now());
+        if (!$request->has('customer_id')) {
+            $customers = $paymentsQuery->get()
+                ->pluck('customer')
+                ->unique('id')
+                ->values();
+            return response()->json(['customers' => $customers]);
+        }
+
+        $connections = $paymentsQuery->whereHas('customer', function($customerQuery) use ($request) {
+            $customerQuery->where('id', $request->customer_id);
         })
-        ->get()
-        ->map(function ($connection) {
-            return [
-                'id' => $connection->id,
-                'name' => $connection->name,
-            ];
-        });
+            ->get()
+            ->pluck('debt.waterConnection')
+            ->unique('id')
+            ->values();
 
-    return response()->json(['waterConnections' => $waterConnections]);
+        return response()->json(['waterConnections' => $connections]);
     }
-
 
     public function getAdvanceDebtDates(Request $request)
     {
-        $customerId = $request->input('customer_id');
         $connectionId = $request->input('water_connection_id');
 
         Log::info('getAdvanceDebtDates', [
-            'customer_id' => $customerId,
             'water_connection_id' => $connectionId
         ]);
 
@@ -102,20 +87,14 @@ class AdvancePaymentController extends Controller
 
         $debt = DB::table('debts')
             ->join('payments', 'debts.id', '=', 'payments.debt_id')
-            ->join('water_connections', 'debts.water_connection_id', '=', 'water_connections.id')
             ->where('debts.status', 'paid')
-            ->where('water_connections.customer_id', $customerId)
-            ->where('debts.water_connection_id', $connectionId)
+            ->where('debts.water_connection_id', $connectionId) 
             ->where('debts.end_date', '>', $now)
             ->orderBy('debts.start_date', 'asc')
             ->select('debts.start_date', 'debts.end_date')
             ->first();
 
-        if ($debt) {
-            return response()->json([
-                'start_date' => $debt->start_date,
-                'end_date' => $debt->end_date,
-            ]);
-        } 
+        return $debt 
+            ? response()->json(['start_date' => $debt->start_date,'end_date' => $debt->end_date,]) : response();
     }
 }
